refactor: extract shared gh_utils module to consolidate duplicated helpers
