+++ run_merges.py
@@ -1,56 +1,23 @@
 import json
-import os
 import subprocess
 import time
-from functools import lru_cache
 
-
-def _parse_env_line(line, env_dict):
-    line = line.strip()
-    if not line:
-        return
-    if line.startswith("#"):
-        return
-    if line.startswith("export "):
-        line = line[7:].strip()
-    if "=" not in line:
-        return
-    key, val = line.split("=", 1)
-    env_dict[key] = val.strip("'\"")
-
-
-@lru_cache(maxsize=None)
-def _get_parsed_env_vars():
-    # ⚡ Bolt Optimization: Cache only the parsed variables from the file to prevent redundant IO reads, while keeping it safe from mutable dictionary cache poisoning
-    parsed_vars = {}
-    env_path = os.getenv("GH_TOKEN_ENV_PATH", "../email-security-pipeline/GH_TOKEN.env")
-    try:
-        with open(env_path, "r") as f:
-            for line in f:
-                _parse_env_line(line, parsed_vars)
-    except FileNotFoundError:
-        pass
-    return parsed_vars
-
-def _load_gh_token_env():
-    env = os.environ.copy()
-    env.update(_get_parsed_env_vars())
-    return env
+from gh_utils import load_gh_token_env
 
 
 def run_gh(cmd_list):
-    env = _load_gh_token_env()
+    env = load_gh_token_env()
     result = subprocess.run(cmd_list, capture_output=True, text=True, env=env)
     if result.returncode != 0:
         return None
     try:
         return json.loads(result.stdout)
-    except:
+    except (ValueError, json.JSONDecodeError):
         return result.stdout
 
 
 def get_diff(repo, pr):
-    env = _load_gh_token_env()
+    env = load_gh_token_env()
     result = subprocess.run(
         ["gh", "pr", "diff", str(pr), "-R", str(repo)],
         capture_output=True,
@@ -221,7 +188,7 @@ for repo, pr, title in queue:
         continue
 
     print(f"Gate 2 passed. Merging...")
-    env = _load_gh_token_env()
+    env = load_gh_token_env()
     res = subprocess.run(
         ["gh", "pr", "merge", str(pr), "-R", str(repo), "--squash", "--delete-branch"],
         capture_output=True,
